% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_minion_out.R
\name{get_minion_out}
\alias{get_minion_out}
\title{Generate a keyword parser for Minion search output.}
\usage{
get_minion_out(key = "Solutions Found")
}
\arguments{
\item{key}{Output listing value to parse}
}
\description{
Generate a keyword parser for Minion search output.
}

