% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minion.R
\name{minion}
\alias{minion}
\title{Run the minion program on the input file whose path is given.}
\usage{
minion(x)
}
\arguments{
\item{input_file_path}{Minion file to process}
}
\description{
Run the minion program on the input file whose path is given.
}

