% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minionProblem.R
\docType{data}
\name{minionProblem}
\alias{minionProblem}
\title{Base class for representation of Minion problems.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
minionProblem
}
\arguments{
\item{input}{}
}
\description{
Base class for representation of Minion problems.
}
\keyword{datasets}

