% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minionProblem.R
\name{print.MinionProblem}
\alias{print.MinionProblem}
\title{Print method for MinionProblem objects.}
\usage{
\method{print}{MinionProblem}(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Print method for MinionProblem objects.
}

