% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rminion-package.R
\docType{package}
\name{rminion-package}
\alias{rminion}
\alias{rminion-package}
\title{rminion: Use the Minion Constraint Solver from R}
\description{
Use the Minion constraint solver to solve constraint satisfaction problems.
}
\author{
\strong{Maintainer}: Matthew Henderson \email{matthew.james.henderson@gmail.com} (\href{https://orcid.org/0000-0001-7949-8208}{ORCID})

}
\keyword{internal}
